requirements.txt
Flask
Redis
pillow
cassandra.driver
tensorflow
matplotlib
opencv-python

